chore(tests): organize device tests into folders
