+++ tests/device_tests/binance/test_get_address.py
