+++ tests/device_tests/binance/test_get_public_key.py
