+++ tests/device_tests/binance/test_sign_tx.py
