+++ tests/device_tests/monero/test_getaddress.py
@@ -16,26 +16,27 @@
 
 import pytest
 
-from trezorlib import nem
+from trezorlib import monero
 from trezorlib.tools import parse_path
 
-from ..common import MNEMONIC12
+from ...common import MNEMONIC12
 
 
 @pytest.mark.altcoin
-@pytest.mark.nem
-class TestMsgNEMGetaddress:
+@pytest.mark.monero
+@pytest.mark.skip_t1
+class TestMsgMoneroGetaddress:
     @pytest.mark.setup_client(mnemonic=MNEMONIC12)
-    def test_nem_getaddress(self, client):
+    def test_monero_getaddress(self, client):
         assert (
-            nem.get_address(
-                client, parse_path("m/44'/1'/0'/0'/0'"), 0x68, show_display=True
-            )
-            == "NB3JCHVARQNGDS3UVGAJPTFE22UQFGMCQGHUBWQN"
+            monero.get_address(client, parse_path("m/44h/128h/0h"), show_display=True)
+            == b"4Ahp23WfMrMFK3wYL2hLWQFGt87ZTeRkufS6JoQZu6MEFDokAQeGWmu9MA3GFq1yVLSJQbKJqVAn9F9DLYGpRzRAEXqAXKM"
         )
         assert (
-            nem.get_address(
-                client, parse_path("m/44'/1'/0'/0'/0'"), 0x98, show_display=True
-            )
-            == "TB3JCHVARQNGDS3UVGAJPTFE22UQFGMCQHSBNBMF"
+            monero.get_address(client, parse_path("m/44h/128h/1h"), show_display=True)
+            == b"44iAazhoAkv5a5RqLNVyh82a1n3ceNggmN4Ho7bUBJ14WkEVR8uFTe9f7v5rNnJ2kEbVXxfXiRzsD5Jtc6NvBi4D6WNHPie"
+        )
+        assert (
+            monero.get_address(client, parse_path("m/44h/128h/2h"), show_display=True)
+            == b"47ejhmbZ4wHUhXaqA4b7PN667oPMkokf4ZkNdWrMSPy9TNaLVr7vLqVUQHh2MnmaAEiyrvLsX8xUf99q3j1iAeMV8YvSFcH"
         )
