+++ tests/device_tests/monero/test_getwatchkey.py
@@ -19,7 +19,7 @@ import pytest
 from trezorlib import monero
 from trezorlib.tools import parse_path
 
-from ..common import MNEMONIC12
+from ...common import MNEMONIC12
 
 
 @pytest.mark.altcoin
