+++ tests/device_tests/recovery/test_bip39.py
@@ -18,7 +18,7 @@ import pytest
 
 from trezorlib import device, messages as proto
 
-from ..common import MNEMONIC12
+from ...common import MNEMONIC12
 
 PIN4 = "1234"
 PIN6 = "789456"
