+++ tests/device_tests/recovery/test_bip39_t2.py
@@ -18,7 +18,7 @@ import pytest
 
 from trezorlib import device, exceptions, messages
 
-from ..common import MNEMONIC12
+from ...common import MNEMONIC12
 
 pytestmark = pytest.mark.skip_t1
