+++ tests/device_tests/recovery/test_slip39_advanced_dryrun.py
@@ -19,23 +19,24 @@ import pytest
 from trezorlib import device, messages
 from trezorlib.exceptions import TrezorFailure
 
-from ..common import recovery_enter_shares
+from ...common import MNEMONIC_SLIP39_ADVANCED_20, recovery_enter_shares
 
 pytestmark = pytest.mark.skip_t1
 
-SHARES_20_2of3 = [
-    "crush merchant academic acid dream decision orbit smug trend trust painting slice glad crunch veteran lunch friar satoshi engage aquatic",
-    "crush merchant academic agency devote eyebrow disaster island deploy flip toxic budget numerous airport loyalty fitness resident learn sympathy daughter",
-    "crush merchant academic always course verdict rescue paces fridge museum energy solution space ladybug junction national biology game fawn coal",
+INVALID_SHARES_SLIP39_ADVANCED_20 = [
+    "chest garlic acrobat leaf diploma thank soul predator grant laundry camera license language likely slim twice amount rich total carve",
+    "chest garlic acrobat lily adequate dwarf genius wolf faint nylon scroll national necklace leader pants literary lift axle watch midst",
+    "chest garlic beard leaf coastal album dramatic learn identify angry dismiss goat plan describe round writing primary surprise sprinkle orbit",
+    "chest garlic beard lily burden pistol retreat pickup emphasis large gesture hand eyebrow season pleasure genuine election skunk champion income",
 ]
 
-INVALID_SHARES_20_2of3 = [
-    "gesture necklace academic acid civil round fiber buyer swing ancient jerky kitchen chest dining enjoy tension museum increase various rebuild",
-    "gesture necklace academic agency decrease justice ounce dragon shaped unknown material answer dress wrote smell family squeeze diet angry husband",
+# Extra share from another group to make sure it does not matter.
+EXTRA_GROUP_SHARE = [
+    "eraser senior decision smug corner ruin rescue cubic angel tackle skin skunk program roster trash rumor slush angel flea amazing"
 ]
 
 
-@pytest.mark.setup_client(mnemonic=SHARES_20_2of3[0:2])
+@pytest.mark.setup_client(mnemonic=MNEMONIC_SLIP39_ADVANCED_20, passphrase=False)
 def test_2of3_dryrun(client):
     debug = client.debug
 
@@ -43,7 +44,9 @@ def test_2of3_dryrun(client):
         yield  # Confirm Dryrun
         debug.press_yes()
         # run recovery flow
-        yield from recovery_enter_shares(debug, SHARES_20_2of3[1:3])
+        yield from recovery_enter_shares(
+            debug, EXTRA_GROUP_SHARE + MNEMONIC_SLIP39_ADVANCED_20, groups=True
+        )
 
     with client:
         client.set_input_flow(input_flow)
@@ -62,7 +65,7 @@ def test_2of3_dryrun(client):
     )
 
 
-@pytest.mark.setup_client(mnemonic=SHARES_20_2of3[0:2])
+@pytest.mark.setup_client(mnemonic=MNEMONIC_SLIP39_ADVANCED_20)
 def test_2of3_invalid_seed_dryrun(client):
     debug = client.debug
 
@@ -70,7 +73,9 @@ def test_2of3_invalid_seed_dryrun(client):
         yield  # Confirm Dryrun
         debug.press_yes()
         # run recovery flow
-        yield from recovery_enter_shares(debug, INVALID_SHARES_20_2of3)
+        yield from recovery_enter_shares(
+            debug, INVALID_SHARES_SLIP39_ADVANCED_20, groups=True
+        )
 
     # test fails because of different seed on device
     with client, pytest.raises(
