+++ tests/device_tests/reset/test_bip39.py
@@ -19,7 +19,7 @@ from mnemonic import Mnemonic
 
 from trezorlib import device, messages as proto
 
-from ..common import generate_entropy
+from ...common import generate_entropy
 
 
 def reset_device(client, strength):
