+++ tests/device_tests/reset/test_slip39_advanced.py
@@ -23,7 +23,7 @@ from trezorlib import device, messages as proto
 from trezorlib.exceptions import TrezorFailure
 from trezorlib.messages import BackupType, ButtonRequestType as B
 
-from ..common import click_through, generate_entropy, read_and_confirm_mnemonic
+from ...common import click_through, generate_entropy, read_and_confirm_mnemonic
 
 EXTERNAL_ENTROPY = b"zlutoucky kun upel divoke ody" * 2
