+++ tests/device_tests/ripple/test_get_address.py
