+++ tests/device_tests/tezos/test_getaddress.py
