+++ tests/device_tests/tezos/test_getpublickey.py
