+++ tests/device_tests/tezos/test_sign_tx.py
